+++ packages/apps/quality/src/components/atoms/ControlEntryCard/ControlEntryCard.tsx
@@ -20,9 +20,9 @@ import React, {useEffect, useMemo, useState} from 'react';
 import {StyleSheet, TouchableOpacity, View} from 'react-native';
 import {Card, ProgressBar, Text, useThemeColor} from '@axelor/aos-mobile-ui';
 import {useTranslator, useDispatch} from '@axelor/aos-mobile-core';
-import {DateDisplay} from '../atoms';
-import {ControlEntry} from '../../types';
-import {searchControlEntrySampleApi} from '../../api';
+import {ControlEntry} from '../../../types';
+import {searchControlEntrySampleApi} from '../../../api';
+import {DateDisplay} from '..';
 
 interface ControlEntryCardProps {
   style?: any;
