+++ packages/apps/quality/src/components/index.js
@@ -16,4 +16,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export {default as controlEntryCard} from './ControlEntryCard/ControlEntryCard';
+export * from './atoms';
