+++ packages/apps/quality/src/i18n/en.json
@@ -2,5 +2,9 @@
   "Quality_Quality": "Quality",
   "Quality_ControlEntry": "Control Entry",
   "Quality_Sample": "Sample",
+  "Quality_Status_Draft": "Draft",
+  "Quality_Status_InProgress": "In progress",
+  "Quality_Status_Completed": "Completed",
+  "Quality_Status_Canceled": "Canceled",
   "Quality_SliceAction_SearchControlEntry": "search control entry"
 }
