+++ packages/apps/quality/src/models/sortFields.ts
@@ -20,4 +20,5 @@ import {SortFields} from '@axelor/aos-mobile-core';
 
 export const quality_sortFields: SortFields = {
   quality_controlEntry: ['entryDateTime', 'createdOn'],
+  quality_controlEntrySample: ['entrySampleNbr'],
 };
